adding webpack for project && making module structure for project
