+++ src/components/content/loadButtonTemplate.js
@@ -1,0 +1,3 @@
+export const loadButtonTemplate = `
+<button class="load-more" type="button">load more</button>
+`;
